requirements.txt
django
gunicorn
django-heroku
requests
python-telegram-bot==12.7